Specify serial queue in init for multithreaded apps
Previously, the code would use .global(qos: .userInitiated) to
specify which queue to use. This is a concurrent queue, so many
quick calls to functions in SwiftOCR can result in EXC_BAD_ACCESS
and other memory errors.

This does not change the behavior of SwiftOCR for users who do not
specifically opt into the fix by specifying a serial queue.

Specify a serial queue like the following in your client code to
opt in:

     let ocrQueue = DispatchQueue(label: "com.yourdomain.swiftocr")
     let ocr = SwiftOCR(dispatchQueue: ocrQueue)